--- models/custom-view-installation/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@commercetools-test-data/custom-view-installation",
-  "version": "6.4.1",
-  "description": "Data model for internal Merchan Center API Custom View Installation",
-  "bugs": "https://github.com/commercetools/test-data/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/commercetools/test-data.git",
-    "directory": "models/custom-view-installation"
-  },
-  "keywords": ["javascript", "typescript", "test-data"],
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "dist/commercetools-test-data-custom-view-installation.cjs.js",
-  "module": "dist/commercetools-test-data-custom-view-installation.esm.js",
-  "files": ["dist", "package.json", "LICENSE", "README.md"],
-  "dependencies": {
-    "@babel/runtime": "^7.17.9",
-    "@babel/runtime-corejs3": "^7.17.9",
-    "@commercetools-frontend/application-config": "^22.10.0",
-    "@commercetools-frontend/constants": "^22.10.0",
-    "@commercetools-test-data/commons": "6.4.1",
-    "@commercetools-test-data/custom-view": "6.4.1",
-    "@commercetools-test-data/core": "6.4.1",
-    "@commercetools-test-data/utils": "6.4.1",
-    "@commercetools/platform-sdk": "^6.0.0",
-    "@faker-js/faker": "^8.0.0",
-    "@types/lodash": "^4.14.182",
-    "lodash": "^4.17.21"
-  }
-}
