--- models/custom-view-installation/src/index.ts
@@ -1,7 +1,0 @@
-// Export types
-export * from './restricted-custom-view-installation-for-organization/types';
-export * from './custom-view-installation-permission/types';
-
-// Export models
-export * as RestrictedCustomViewInstallationForOrganization from './restricted-custom-view-installation-for-organization';
-export * as CustomViewInstallationPermission from './custom-view-installation-permission';
