+++ models/custom-view/src/custom-view-installation-permission/presets/manage-only-permissions.ts
@@ -1,9 +1,9 @@
 import { entryPointUriPathToResourceAccesses } from '@commercetools-frontend/application-config/ssr';
 import { CUSTOM_VIEW_HOST_ENTRY_POINT_URI_PATH } from '@commercetools-frontend/constants';
-import { supportedManageOAuthScopes } from '@commercetools-test-data/custom-view';
 import camelCase from 'lodash/camelCase';
 import sampleSize from 'lodash/sampleSize';
 import upperFirst from 'lodash/upperFirst';
+import { supportedManageOAuthScopes } from '../../custom-view-permission/constants';
 import CustomViewInstallationPermission from '../builder';
 
 const preset = (additionalPermission = '') => {
