+++ models/custom-view/src/custom-view-installation-permission/transformers.ts
@@ -5,6 +5,10 @@ import type {
 } from './types';
 
 const transformers = {
+  default: Transformer<
+    TCustomViewInstallationPermission,
+    TCustomViewInstallationPermission
+  >('default', {}),
   graphql: Transformer<
     TCustomViewInstallationPermission,
     TCustomViewInstallationPermissionGraphql
