+++ models/custom-view/src/custom-view-installation/index.ts
@@ -1,0 +1,3 @@
+// Export models
+export * as RestrictedCustomViewInstallationForOrganization from './restricted-custom-view-installation-for-organization';
+export * as CustomViewInstallationPermission from '../custom-view-installation-permission';
