+++ models/custom-view/src/custom-view-installation/types.ts
@@ -1,0 +1,2 @@
+// Export types
+export * from './restricted-custom-view-installation-for-organization/types';
