+++ models/custom-view/src/index.ts
@@ -2,6 +2,8 @@
 export * from './custom-view/types';
 export * from './custom-view-permission/types';
 export * from './custom-view-type-settings-for-custom-panel/types';
+export * from './custom-view-installation/types';
+export * from './custom-view-installation-permission/types';
 
 // Export constants
 export * from './custom-view/constants';
@@ -15,3 +17,5 @@ export * as CustomViewPermission from './custom-view-permission';
 export * as CustomViewPermissionDraft from './custom-view-permission-draft';
 
 export * as CustomViewTypeSettingsForCustomPanel from './custom-view-type-settings-for-custom-panel';
+export { RestrictedCustomViewInstallationForOrganization } from './custom-view-installation';
+export * as CustomViewInstallationPermission from './custom-view-installation-permission';
